+++ servers/zts/src/main/java/com/yahoo/athenz/zts/cert/impl/crypki/AbstractHttpCertSigner.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.yahoo.athenz.zts.cert.impl.v3;
+package com.yahoo.athenz.zts.cert.impl.crypki;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -49,6 +49,7 @@ import com.yahoo.athenz.zts.ZTSConsts;
 import com.yahoo.athenz.zts.utils.ZTSUtils;
 
 public abstract class AbstractHttpCertSigner implements CertSigner {
+
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractHttpCertSigner.class);
     private static final String CONTENT_JSON = "application/json";
     
@@ -65,7 +66,6 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
     int requestRetryCount;
     int maxCertExpiryTimeMins;
 
-    
     public AbstractHttpCertSigner() {
 
         PrivateKeyStore privateKeyStore = loadServicePrivateKey();
@@ -100,59 +100,58 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
         
         x509CertUri = getX509CertUri(serverBaseUri);
         this.connManager = createConnectionPooling(sslContextFactory.getSslContext());
-        this.httpClient = createHttpClient(connectionTimeoutSec, readTimeoutSec, sslContextFactory.getSslContext(), this.connManager);
+        this.httpClient = createHttpClient(connectionTimeoutSec, readTimeoutSec,
+                sslContextFactory.getSslContext(), this.connManager);
 
-        LOGGER.info("HttpCertSigner initialized with url: {} connectionTimeoutSec: {}, readTimeoutSec: {}", x509CertUri, connectionTimeoutSec, readTimeoutSec);
+        LOGGER.info("HttpCertSigner initialized with url: {} connectionTimeoutSec: {}, readTimeoutSec: {}",
+                x509CertUri, connectionTimeoutSec, readTimeoutSec);
         LOGGER.info("HttpCertSigner connection pool stats {} ", this.connManager.getTotalStats().toString());
     }
 
     /**
-     * 
-     * @return
+     * Return x.509 certificate uri based on server uri
+     * @return uri
      */
     public abstract String getX509CertUri(String serverBaseUri);
 
-
     /**
-     * 
-     * @return
+     * Return object based on given csr, usage and expiry
+     * @return CSR Object
      */
     public abstract Object getX509CertSigningRequest(String csr, String keyUsage, int expireMins);
 
     /**
-     * 
-     * @param response
-     * @return
+     * Parse the response from certificate sisnger
+     * @param response input stream
+     * @return response as string
      */
     public abstract String parseResponse(InputStream response) throws IOException;
-    
-    
+
     /**
-     * 
-     * @param sslContext
-     * @return
+     * Create a http client connection manager based on given ssl context
+     * @param sslContext ssl context containing keystore with client key/cert
+     * @return connection manager object
      */
     PoolingHttpClientConnectionManager createConnectionPooling(SSLContext sslContext) {
-        SSLConnectionSocketFactory sslsf = null;
-        Registry<ConnectionSocketFactory> registry = null;
-        PoolingHttpClientConnectionManager poolingHttpClientConnectionManager = null;
-        sslsf = new SSLConnectionSocketFactory(sslContext);
-        registry = RegistryBuilder.<ConnectionSocketFactory>create().register("https", sslsf).build();
-        poolingHttpClientConnectionManager = new PoolingHttpClientConnectionManager(registry);
+
+        SSLConnectionSocketFactory sslsf = new SSLConnectionSocketFactory(sslContext);
+        Registry<ConnectionSocketFactory> registry = RegistryBuilder.<ConnectionSocketFactory>create().register("https", sslsf).build();
+        PoolingHttpClientConnectionManager poolingHttpClientConnectionManager = new PoolingHttpClientConnectionManager(registry);
         
         //route is host + port.  Since we have only one, set the max and the route the same
+
         poolingHttpClientConnectionManager.setDefaultMaxPerRoute(DEFAULT_MAX_POOL_PER_ROUTE);
         poolingHttpClientConnectionManager.setMaxTotal(DEFAULT_MAX_POOL_TOTAL);
         return poolingHttpClientConnectionManager;
     }
 
     /**
-     * 
-     * @param connectionTimeoutSec
-     * @param readTimeoutSec
-     * @param sslContext
-     * @param poolingHttpClientConnectionManager
-     * @return
+     * Create an http client based on given configuration settings
+     * @param connectionTimeoutSec connection timeout in seconds
+     * @param readTimeoutSec read timeout in seconds
+     * @param sslContext ssl context object
+     * @param poolingHttpClientConnectionManager http connection manager object
+     * @return http client
      */
     CloseableHttpClient createHttpClient(int connectionTimeoutSec, int readTimeoutSec, SSLContext sslContext, PoolingHttpClientConnectionManager poolingHttpClientConnectionManager) {
         
@@ -180,7 +179,7 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
             this.sslContextFactory.stop();
             this.httpClient.close();
             if (LOGGER.isDebugEnabled()) {
-                LOGGER.debug("connManager stats close(): {}" , this.connManager.getTotalStats().toString());
+                LOGGER.debug("connManager stats close(): {}", this.connManager.getTotalStats().toString());
             }
             this.connManager.close();
         } catch (Exception ignored) {
@@ -199,8 +198,8 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
         try {
             pkeyFactory = (PrivateKeyStoreFactory) Class.forName(pkeyFactoryClass).newInstance();
         } catch (InstantiationException | IllegalAccessException | ClassNotFoundException e) {
-            LOGGER.error("Invalid PrivateKeyStoreFactory class: " + pkeyFactoryClass
-                    + " error: " + e.getMessage());
+            LOGGER.error("Invalid PrivateKeyStoreFactory class: {} error: {}",
+                    pkeyFactoryClass, e.getMessage());
             throw new IllegalArgumentException("Invalid private key store");
         }
         return pkeyFactory.create();
@@ -208,7 +207,8 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
 
     @Override
     public String generateX509Certificate(String csr, String keyUsage, int expireMins) {
-        StringEntity entity = null;
+
+        StringEntity entity;
         try {
             String requestContent = JACKSON_MAPPER.writeValueAsString(getX509CertSigningRequest(csr, keyUsage, expireMins));
             entity = new StringEntity(requestContent);
@@ -222,12 +222,14 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
         httpPost.setHeader("Content-Type", CONTENT_JSON);
         httpPost.setEntity(entity);
         
-        //Retry
+        // Retry configured number of times before returning failure
+
         for (int i = 0; i < requestRetryCount; i++) {
             try {
                 return processHttpResponse(httpPost, 201);
             } catch (IOException e) {
-                LOGGER.error("Unable to process x509 certificate request to url " + x509CertUri + " Retring " + i + 1 + "/" + requestRetryCount, e);
+                LOGGER.error("Unable to process x509 certificate request to url {}, retrying {}/{}, {}",
+                        x509CertUri, i + 1, requestRetryCount, e);
             }
         }
         
@@ -236,14 +238,15 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
     
 
     /**
-     * 
-     * @param request
-     * @param expectedStatusCode
+     * Process http response from crypki server
+     * @param request http request object
+     * @param expectedStatusCode expected http status code
      * @return x509 Certificate or Null if expectedStatusCode doesn't match or empty response from the server.
      * @throws ClientProtocolException
      * @throws IOException
      */
     String processHttpResponse(HttpUriRequest request, int expectedStatusCode) throws ClientProtocolException, IOException {
+
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("connManager stats before: {}" , this.connManager.getTotalStats().toString());
         }
@@ -255,13 +258,13 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
         int statusCode = response.getStatusLine().getStatusCode();
         if (statusCode != expectedStatusCode) {
             // we have a response but not 201. Don't bother retry
-            LOGGER.error("unable to fetch requested uri '" + x509CertUri + "' status: " + statusCode);
+            LOGGER.error("unable to fetch requested uri '{}' status: {}", x509CertUri, statusCode);
             return null;
         }
         // check for content
         try (InputStream data = response.getEntity().getContent()) {
             if (data == null) {
-                LOGGER.error("received empty response from uri '" + x509CertUri + "' status: " + statusCode);
+                LOGGER.error("received empty response from uri '{}', status:  {}", x509CertUri, statusCode);
                 return null;
             }
             return parseResponse(data);
@@ -278,7 +281,7 @@ public abstract class AbstractHttpCertSigner implements CertSigner {
             LOGGER.error("Unable to process x509 CA certificate request", e);
         }
         if (LOGGER.isDebugEnabled()) {
-            LOGGER.debug("getCACertificate: CA Certificate" + data);
+            LOGGER.debug("getCACertificate: CA Certificate {}", data);
         }
         return data;
     }
